Added some type checking

+++ DiploGM/parse_edit_state.py
@@ -44,7 +44,7 @@ logger = logging.getLogger(__name__)
 manager = Manager()
 
 
-def parse_edit_state(message: str, board: Board) -> dict[str, ...]:
+def parse_edit_state(message: str, board: Board) -> dict[str, str | bytes | int | None]:
     invalid: list[tuple[str, Exception]] = []
     commands = str.splitlines(message)
     for command in commands:
@@ -142,7 +142,7 @@ def _set_phase(keywords: list[str], board: Board) -> None:
     old_turn = board.turn.get_indexed_name()
     new_turn = parse_season(keywords, board.turn.year)
     if new_turn is None:
-        raise ValueError(f"{keywords.join(' ')} is not a valid phase name")
+        raise ValueError(f"{' '.join(keywords)} is not a valid phase name")
     board.turn = new_turn
     get_connection().execute_arbitrary_sql(
         "UPDATE boards SET phase=? WHERE board_id=? and phase=?",
@@ -190,6 +190,8 @@ def _set_province_half_core(keywords: list[str], board: Board) -> None:
 
 def _set_player_color(keywords: list[str], board: Board) -> None:
     player = board.get_player(keywords[0])
+    if not player:
+        raise ValueError(f"Unknown player: {keywords[0]}")
     color = keywords[1].lower()
     if not len(color) == 6 or not all(c in string.hexdigits for c in color):
         raise ValueError(f"Unknown hexadecimal color: {color}")
@@ -239,6 +241,8 @@ def _create_unit(keywords: list[str], board: Board) -> None:
         raise ValueError(f"Invalid Unit Type received: {unit_type}")
 
     player = board.get_player(keywords[1])
+    if not player:
+        raise ValueError(f"Unknown player: {keywords[1]}")
     province, coast = board.get_province_and_coast(" ".join(keywords[2:]))
     # if unit_type == UnitType.FLEET and coast is None:
     #     coast_name = f"{province} coast"
@@ -265,10 +269,14 @@ def _create_unit(keywords: list[str], board: Board) -> None:
 def _create_dislodged_unit(keywords: list[str], board: Board) -> None:
     if board.turn.is_retreats():
         unit_type = get_unit_type(keywords[0])
+        if not unit_type:
+            raise ValueError(f"Invalid Unit Type received: {unit_type}")
         player = board.get_player(keywords[1])
+        if not player:
+            raise ValueError(f"Unknown player: {keywords[1]}")
         province, coast = board.get_province_and_coast(keywords[2])
         retreat_options = set(
-            [board.get_province(province_name) for province_name in keywords[3:]]
+            [board.get_province_and_coast(province_name) for province_name in keywords[3:]]
         )
         if not all(retreat_options):
             raise ValueError(
@@ -297,7 +305,7 @@ def _create_dislodged_unit(keywords: list[str], board: Board) -> None:
                     board.board_id,
                     board.turn.get_indexed_name(),
                     unit.province.get_name(coast),
-                    option.name,
+                    option[0].get_name(option[1]),
                 )
                 for option in retreat_options
             ],
@@ -309,6 +317,8 @@ def _create_dislodged_unit(keywords: list[str], board: Board) -> None:
 def _delete_unit(keywords: list[str], board: Board) -> None:
     province = board.get_province(keywords[0])
     unit = board.delete_unit(province)
+    if not unit:
+        raise RuntimeError(f"No unit to delete in {province}")
     get_connection().execute_arbitrary_sql(
         "DELETE FROM units WHERE board_id=? and phase=? and location=? and is_dislodged=?",
         (
@@ -323,19 +333,23 @@ def _delete_unit(keywords: list[str], board: Board) -> None:
 def _delete_dislodged_unit(keywords: list[str], board: Board) -> None:
     province = board.get_province(keywords[0])
     unit = board.delete_dislodged_unit(province)
+    if not unit:
+        raise RuntimeError(f"No dislodged unit to delete in {province}")
     get_connection().execute_arbitrary_sql(
         "DELETE FROM units WHERE board_id=? and phase=? and location=? and is_dislodged=?",
         (board.board_id, board.turn.get_indexed_name(), unit.province.get_name(unit.coast), True),
     )
     get_connection().execute_arbitrary_sql(
         "DELETE FROM retreat_options WHERE board_id=? and phase=? and origin=?",
-        (board.board_id, board.turn.get_indexed_name(), unit.province.get_name(unit.coast).name),
+        (board.board_id, board.turn.get_indexed_name(), unit.province.get_name(unit.coast)),
     )
 
 
 def _move_unit(keywords: list[str], board: Board) -> None:
     old_province = board.get_province(keywords[0])
     unit = old_province.unit
+    if not unit:
+        raise RuntimeError(f"No unit to move in {old_province}")
     new_province, new_coast = board.get_province_and_coast(keywords[1])
     board.move_unit(unit, new_province, new_coast)
     get_connection().execute_arbitrary_sql(
@@ -402,6 +416,8 @@ def _make_units_claim_provinces(keywords: list[str], board: Board) -> None:
 
 def _set_player_points(keywords: list[str], board: Board) -> None:
     player = board.get_player(keywords[0])
+    if not player:
+        raise ValueError("Unknown player specified")
     points = int(keywords[1])
     if points < 0:
         raise ValueError("Can't have a negative number of points!")
@@ -416,6 +432,8 @@ def _set_player_points(keywords: list[str], board: Board) -> None:
 def _set_player_vassal(keywords: list[str], board: Board) -> None:
     liege = board.get_player(keywords[0])
     vassal = board.get_player(keywords[1])
+    if not liege or not vassal:
+        raise ValueError("Unknown player specified")
     vassal.liege = liege
     liege.vassals.append(vassal)
     get_connection().execute_arbitrary_sql(
@@ -427,6 +445,8 @@ def _set_player_vassal(keywords: list[str], board: Board) -> None:
 def _remove_player_vassal(keywords: list[str], board: Board) -> None:
     player1 = board.get_player(keywords[0])
     player2 = board.get_player(keywords[1])
+    if not player1 or not player2:
+        raise ValueError("Unknown player specified")
     for vassal, liege in ((player1, player2), (player2, player1)):
         if vassal.liege == liege:
             vassal.liege = None
